Rename wikiDiff to diffTool to make it more generic
